+++ backend/src/main/java/pl/edu/ug/inf/lexica/api/UserController.java
@@ -23,7 +23,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}/team")
-    public List<Team> getTeams(@PathVariable Integer id) {
+    public List<Team> getTeams(@PathVariable Long id) {
         return userService.get(id).map(user -> Stream
                 .concat(user.getLeading().stream(), user.getTeams().stream())
                 .map(Team::withSomeInfo)
@@ -32,12 +32,12 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable Integer id) {
+    public User getUser(@PathVariable Long id) {
         return userService.get(id).map(User::withMoreInfo).orElse(new User());
     }
 
     @PostMapping("/{id}/progress")
-    public void addProgress(@RequestBody Progress progress, @PathVariable Integer id) {
+    public void addProgress(@RequestBody Progress progress, @PathVariable Long id) {
         userService.get(id).ifPresent(user -> {
             user.getProgress().add(progress);
             userService.update(user);
@@ -45,7 +45,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}/progress")
-    public List<Progress> getProgress(@PathVariable Integer id) {
+    public List<Progress> getProgress(@PathVariable Long id) {
         return userService.get(id).orElse(new User()).getProgress();
     }
 
@@ -55,7 +55,7 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public void updateUser(@RequestBody User updated, @PathVariable Integer id) {
+    public void updateUser(@RequestBody User updated, @PathVariable Long id) {
         userService.get(id).ifPresent(user -> {
             user.setFirstname(updated.getFirstname());
             user.setSurname(updated.getSurname());
@@ -66,7 +66,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Integer id) {
+    public void deleteUser(@PathVariable Long id) {
         userService.remove(id);
     }
 }
