+++ backend/src/main/java/pl/edu/ug/inf/lexica/repository/TaskTypeRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import pl.edu.ug.inf.lexica.domain.TaskType;
 
 @Repository
-public interface TaskTypeRepository extends JpaRepository<TaskType, Integer> { }
+public interface TaskTypeRepository extends JpaRepository<TaskType, Long> { }
