+++ backend/src/main/java/pl/edu/ug/inf/lexica/repository/UserRepository.java
@@ -1,17 +1,14 @@
 package pl.edu.ug.inf.lexica.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import pl.edu.ug.inf.lexica.domain.User;
 
-import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findById(Integer id);
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findById(Long id);
     List<User> findAll();
 }
